+++ CHANGELOG.rst
@@ -43,6 +43,8 @@ Change Log
   if numba is available but you don't want it)
 - [ADDED] the method `act.decompose_as_unary_actions(...)` to automatically
   decompose a "complex" action on its unary counterpart. 
+- [ADDED] the env attribute `env._reward_to_obs` that allows to pass information to the observation directly
+  from the reward (this can only be used by regular environment and not by `obs.simulate` nor by `ForecastEnv`)
 - [IMPROVED] the method `act.as_serializable_dict()` to work better when exporting / importing actions on different 
   grids (the output dictionary for `set_bus` and `change_bus` now split the keys between all elements types 
   instead of relying on the "topo_vect" order (which might vary))